special handling for xvfb GL-CI /tmp/.X11-unix by run.sh

+++ run.sh
@@ -1,9 +1,15 @@
-#!/bin/bash
+#!/bin/sh
 
 XSOCK=/tmp/.X11-unix
-XAUTH=/tmp/.docker.xauth
-touch $XAUTH
-xauth nlist $DISPLAY < /dev/null | sed -e 's/^..../ffff/' | xauth -f $XAUTH nmerge -
+if [ -z ${XVFBtmp+x} ]; then 
+    XAUTH=/tmp/.docker.xauth
+    touch $XAUTH
+    xauth nlist $DISPLAY < /dev/null | sed -e 's/^..../ffff/' | xauth -f $XAUTH nmerge -
+
+    DOCKERARGS="--volume $XSOCK:$XSOCK:rw --volume $XAUTH:$XAUTH:rw --env XAUTHORITY=$XAUTH"
+else
+    DOCKERARGS="--volume $XVFBtmp:$XSOCK:rw"
+fi
 
 mkdir -p $HOME/.civctp2/save/
 mkdir -p $HOME/.civctp2/logs/ # if not existent created by docker then owned by root
@@ -18,14 +24,11 @@ fi
 ## ./run.sh -v $HOME/ctp2CD/ctp2_program/ctp/music/:/opt/ctp2/ctp2_program/ctp/music/:ro registry.gitlab.com/civctp2/civctp2/master:latest ./ctp2 fullscreen
 docker run \
        --rm \
-       --volume $XSOCK:$XSOCK:rw \
-       --volume $XAUTH:$XAUTH:rw \
+       $DOCKERARGS \
        --device /dev/dri:/dev/dri \
        --device /dev/snd:/dev/snd \
        --env "ALSA_CARD=0" \
-       --env "XAUTHORITY=${XAUTH}" \
        --env "DISPLAY" \
-       --user "diUser" \
        -v $HOME/.civctp2/userprofile.txt:/opt/ctp2/ctp2_program/ctp/userprofile.txt \
        -v $HOME/.civctp2/userkeymap.txt:/opt/ctp2/ctp2_program/ctp/userkeymap.txt \
        $CON \
@@ -33,4 +36,6 @@ docker run \
        -v $HOME/.civctp2/logs/:/opt/ctp2/ctp2_program/ctp/logs \
        $@
 
-rm $XAUTH # remove to avoid accumulation of xauth settings
+if [ -z ${XVFBtmp+x} ]; then 
+    rm $XAUTH # remove to avoid accumulation of xauth settings
+fi
